cleanup of genesis state
